Satisfy iOS package conventions
