+++ Packages/iOS/CmuxMobileSupport/Sources/CmuxMobileSupport/UITestConfig.swift
@@ -99,19 +99,8 @@ public struct UITestConfig {
     /// debug chat fixture directly so XCUITest can exercise chat layout without
     /// sign-in, pairing, or workspace-shell timing. DEBUG-only.
     public static var agentChatPreviewEnabled: Bool {
-        agentChatPreviewEnabled(from: ProcessInfo.processInfo.environment)
-    }
-
-    /// The standalone agent-chat preview flag for an explicit environment.
-    ///
-    /// - Parameter env: The environment dictionary to read.
-    /// - Returns: `true` when `CMUX_UITEST_AGENT_CHAT_PREVIEW=1` in DEBUG builds.
-    public static func agentChatPreviewEnabled(from env: [String: String]) -> Bool {
-        #if DEBUG
-        return env["CMUX_UITEST_AGENT_CHAT_PREVIEW"] == "1"
-        #else
-        return false
-        #endif
+        UITestEnvironmentConfig(environment: ProcessInfo.processInfo.environment)
+            .agentChatPreviewEnabled
     }
 
     /// Whether mock data is enabled for an explicit environment.
